apps/rtsp-virtual-webcam/requirements.txt
opencv-python>=4.8.0
numpy>=1.20.0
pyvirtualcam>=0.11.0
pystray>=0.19.0
Pillow>=9.0.0
